Move client classes
